compress level doesn't need to be 32 bits wide

+++ include/sys/zio.h
@@ -337,7 +337,7 @@ struct zbookmark_phys {
 typedef struct zio_prop {
 	enum zio_checksum	zp_checksum;
 	enum zio_compress	zp_compress;
-	int32_t			zp_complevel;
+	int8_t			zp_complevel;
 	dmu_object_type_t	zp_type;
 	uint8_t			zp_level;
 	uint8_t			zp_copies;
@@ -629,8 +629,8 @@ extern enum zio_checksum zio_checksum_dedup_select(spa_t *spa,
     enum zio_checksum child, enum zio_checksum parent);
 extern enum zio_compress zio_compress_select(spa_t *spa,
     enum zio_compress child, enum zio_compress parent);
-extern int32_t zio_complevel_select(spa_t *spa, int32_t child,
-    int32_t parent);
+extern int8_t zio_complevel_select(spa_t *spa, int8_t child,
+    int8_t parent);
 
 extern void zio_suspend(spa_t *spa, zio_t *zio, zio_suspend_reason_t);
 extern int zio_resume(spa_t *spa);
